refactor: make auth form component and add auth page
