+++ pages/auth.js
@@ -1,0 +1,9 @@
+import AuthForm from "../components/authForm";
+
+export default function AuthPage() {
+  return (
+    <div className="flex justify-center">
+      <AuthForm />
+    </div>
+  );
+}
